Initial spike attempt.
Create an initial attempt at spiking out the clients service. Creates
the clients and redirect URI types, along with our storer interface.
Also sets up the in-memory storer and writes some tests to make sure
it's working. Spike out the Postgres storer and the API, too.

I decided to SHA-256 hash the client secrets, since they're sensitive
data and I'd like to not have someone be able to authenticate as any
client just by reading the database. But I didn't bother salting or
PBKDFing them, because they're going to be high-entropy values already,
so there's no point. The likelihood of a collision is almost impossible,
so a salt won't do much for us. And the amount of entropy in the secrets
is already enough to make brute forcing their SHA 256 sums prohibitive.