Continue work on storers.
Update AddRedirectURIs and RemoveRedirectURIs methods of the Storer
interface to not require the clientID anymore, because it wasn't used.

Update all the uses of the AddRedirectURIs and RemoveRedirectURIs
methods to take the new function signature into account.

Implement the Postgres storer. It still has a bit of error handling to
do, but passes all the current tests.

Fix the base table name for Postgres tests to be clients_test_, instead
of accounts_test_.

Add the new dependencies to go.mod and go.sum.

Fix the go generate comment in clients.go so our migrations package gets
updated when `go generate .` is called.