Finish moving things around, for now.
Move the Storer type definition from clients.go into its own storer.go.

Move the storer_test.go file from the storers package into the main
package, to avoid having a test-only package. Change its package
declaration to clients_test, to avoid circular imports and only allow
using the exported parts of the clients package.

Move the migrations package to a subpackage of the postgres package, to
group it with the postgres storer implementation it exists for.

Move the sql folder to a subpackage of the postgres package, to group it
with the postgres storer implementation it exists for.

Move the `go generate` command that generates the migrations package
into the postgres package.