Flesh out handlers.
Finish our API handlers. This involved removing the ability to add
redirect URIs when creating clients, because we couldn't do that as a
transaction based on the underlying Storer, and if adding Redirect URIs
failed, we'd get a Client created and return an error to the user, who
would reasonably retry the request, creating another new client. Which
is undesirable. So now clients should be created, then have their
redirect URIs added in a separate request, which sidesteps this problem.