Reformat endpoints.
Reformate endpoints to fit within 80 characters per line, and to avoid
repetition. Also, remove the dependency on the APIv1 type for our
logging middleware, as the log will be read from the context upstream
instead of from the APIv1 type. It is the job of an upstream middleware
to inject a logger.