Make URIs unique, return more helpful errors from postgres.
When inserting our redirect URIs for clients, postgres now considers
them unique. It also will return more helpful errors about which URI
violated which constraint.

This had to be done by regexp parsing the Detail field of the returned
error, which is incredibly gross, but PostgreSQL doesn't actually
surface the information on which row in a multi-row insert triggered the
constraint violation _except_ embedded in the user-facing Detail field,
which is definitely very cool and not at all annoying.