Test failure cases of client storer operations.
Test adding a client with an ID that already exists.

Test updating a client that doesn't exist. Stop returning an
ErrClientNotFound from the memory.Storer, as none of our other packages
return an error when updating something that doesn't exist. Maybe they
should? But one thing at a time.

Test deleting a client that doesn't exist. Stop returning an
ErrClientNotFound from the memory.Storer, as none of our other packages
return an error when removing something that doesn't exist.